+++ lib/preprocessor-configuration.test.ts
@@ -1010,6 +1010,25 @@ describe("resolve()", () => {
             setValueFn,
           });
         });
+
+        describe("addVideos", () => {
+          const getValueFn = (
+            configuration: IPreprocessorConfiguration,
+          ): boolean => configuration.attachments.addVideos;
+
+          const setValueFn = (
+            configuration: IBaseUserConfiguration,
+            value: boolean,
+          ) => (configuration.attachments = { addVideos: value });
+
+          basicBooleanExample({
+            testingType,
+            default: false,
+            environmentKey: "attachmentsAddVideos",
+            getValueFn,
+            setValueFn,
+          });
+        });
       });
     });
   }