+++ lib/preprocessor-configuration.ts
@@ -84,6 +84,7 @@ const EnvironmentOverrides = D.partial({
   omitFiltered: StringishToBoolean,
   dryRun: StringishToBoolean,
   attachmentsAddScreenshots: StringishToBoolean,
+  attachmentsAddVideos: StringishToBoolean,
 });
 
 type IEnvironmentOverrides = D.TypeOf<typeof EnvironmentOverrides>;
@@ -118,6 +119,7 @@ const BaseConfiguration = D.partial({
   dryRun: D.boolean,
   attachments: D.partial({
     addScreenshots: D.boolean,
+    addVideos: D.boolean,
   }),
 });
 
@@ -164,6 +166,7 @@ export interface IPreprocessorConfiguration {
   readonly dryRun: boolean;
   readonly attachments: {
     addScreenshots: boolean;
+    addVideos: boolean;
   };
 }
 
@@ -312,6 +315,11 @@ export function combineIntoConfiguration(
       specific?.attachments?.addScreenshots ??
       unspecific.attachments?.addScreenshots ??
       true,
+    addVideos:
+      overrides.attachmentsAddVideos ??
+      specific?.attachments?.addVideos ??
+      unspecific.attachments?.addVideos ??
+      false,
   };
 
   return {